# openapi.yaml
openapi: 3.0.3
info: { title: Openmind API, version: 0.0.1 }
security:
  - bearerAuth: []
paths:
...
  /items:
    post:
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateItemRequest" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
    get:
      operationId: listItems
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: cursor, in: query, schema: { type: string }, description: "Opaque cursor from a previous page's nextCursor. Omit for the first page." }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemPage" }
        "400": { description: invalid cursor }
  /items/{id}:
    get:
      operationId: getItem
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: item detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemDetail" }
        "404":
          description: not found
    patch:
      operationId: patchItem
      description: "Edit an item. Currently supports replacing the full user-tags list (userTags). Tags are canonicalised server-side (trimmed, lowercased, deduped, capped)."
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateItemRequest" }
      responses:
        "200":
          description: updated item detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemDetail" }
        "400": { description: bad request }
        "404": { description: not found }
    delete:
      operationId: deleteItem
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: deleted }
        "404": { description: not found }
...
  /search:
    get:
      operationId: searchItems
      description: >
        Text search (q), colour-proximity (color), and/or structured filters
        (types, domains), optionally limited by scope. Set parse=true to
        interpret q as a natural-language query — the AI provider may extract
        text, colour, types, and domains, all fused into one search. Explicit
        params win over parsed values. At least one of q, color, types, or
        domains is required (scope alone is not enough). Results rank library
        items (direct saves and kept feed items) ahead of unkept feed-river
        matches when scope includes them; within each group ordering is by
        descending fused score.
      parameters:
        - { name: q, in: query, required: false, schema: { type: string } }
        - { name: color, in: query, required: false, schema: { type: string }, description: "Hex (#RRGGBB) or named colour (e.g. cobalt, terracotta); ranks items by nearest palette colour." }
        - name: types
          in: query
          required: false
          style: form
          explode: true
          description: "Card types to include (filter). Repeatable query param."
          schema:
            type: array
            items: { type: string, enum: [article, product, book, recipe, video, tweet, image, note, quote, repo] }
        - name: domains
          in: query
          required: false
          style: form
          explode: true
          description: "URL hosts to include (filter). Normalised lowercase; www stripped. Subdomains match."
          schema:
            type: array
            items: { type: string, minLength: 1 }
        - name: scope
          in: query
          required: false
          description: "library = saved/kept only; all = include unkept feed items. Defaults to all when omitted on /search."
          schema: { type: string, enum: [library, all], default: all }
        - { name: parse, in: query, required: false, schema: { type: boolean }, description: "Interpret q as a natural-language query, splitting it into text + colour + card-type + domain filters via the AI provider. Falls back to a plain text search when no AI provider is configured." }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResponse" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  schemas:
    CreateItemRequest:
      type: object
      description: "Exactly one of url or note must be provided."
      properties:
        url: { type: string, format: uri }
        note: { type: string, minLength: 1, maxLength: 10000 }
    Item:
      type: object
      required: [id, url, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        title: { type: string }
        summary: { type: string }
        leadImageUrl: { type: string }
        tags: { type: array, items: { type: string } }
        userTags: { type: array, items: { type: string } }
        palette: { type: array, items: { type: string } }
        cardType: { type: string, enum: [article, product, book, recipe, video, tweet, image, note, quote, repo] }
        status: { type: string, enum: [pending, enriched, failed] }
        createdAt: { type: string, format: date-time }
        pinnedAt: { type: string, format: date-time, nullable: true, description: "When the item was pinned to the Desk; null if not pinned." }
        pageCount: { type: integer, nullable: true, description: "Page count when the item's original is a stored PDF; null otherwise." }
        feedId: { type: string, format: uuid, nullable: true, description: "The feed this item originated from; null if not feed-sourced." }
        keptAt: { type: string, format: date-time, nullable: true, description: "When the item was kept in the library from its feed; null if not kept." }
    ItemPage:
      type: object
      description: >
        One page of items, newest first. nextCursor is opaque — pass it back as
        ?cursor= to fetch the following page. It is absent when there are no
        more items, so its presence is the only "has more" signal a client needs.
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Item" }
        nextCursor: { type: string }
    ItemDetail:
      allOf:
        - $ref: "#/components/schemas/Item"
        - type: object
          required: [body]
          properties:
            body: { type: string }
            jevSuggestions:
              $ref: "#/components/schemas/JevSuggestions"
              description: >
                Phase 2 capture chips for this item when AI-assisted organisation
                produced applied/suggested tags. Absent when there is no capture
                decision or nothing left to show.
    JevSuggestions:
      type: object
      required: [action, suggestedTags, appliedTags]
      properties:
        action:
          type: string
          description: "Row-level capture action: applied | suggested | skipped | shadow"
        suggestedTags:
          type: array
          items: { type: string }
          description: Mid-confidence tags not yet accepted, dismissed, or present on the item
        appliedTags:
          type: array
          items: { type: string }
          description: Auto-applied tags still on the item (one-tap undo)
        userVerdict:
          type: string
          nullable: true
          description: kept | changed | removed when the user has reacted
...
    UpdateItemRequest:
      type: object
      description: "Fields to update on an item. Both userTags and pinned are optional; omit both for a no-op edit that is rejected as a bad request."
      properties:
        userTags: { type: array, items: { type: string }, description: "Full replacement user-tags list. An empty array clears all user tags." }
        pinned: { type: boolean, description: "Pin (true) or unpin (false) the item on the Desk. Pinning sets pinnedAt to now; unpinning clears it." }
        kept: { type: boolean, description: "Keep (true) the item in the library independent of its feed. Sets keptAt to now." }
...
    SearchResult:
      type: object
      required: [item, score]
      properties:
        item: { $ref: "#/components/schemas/Item" }
        score: { type: number }
    SearchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items: { $ref: "#/components/schemas/SearchResult" }
        understood:
          $ref: "#/components/schemas/UnderstoodQuery"
    UnderstoodQuery:
      type: object
      description: "How a natural-language query was interpreted (present only when parse=true). Reflects the values actually searched."
      properties:
        text: { type: string, description: "The free-text portion searched." }
        color: { type: string, description: "The colour searched, if any." }
        types:
          type: array
          description: "Card-type filters applied, if any."
          items: { type: string, enum: [article, product, book, recipe, video, tweet, image, note, quote, repo] }
        domains:
          type: array
          description: "URL host filters applied, if any."
          items: { type: string }
...